requirements.txt
python-dotenv
requests
retry
openai